[Refactor] 마일스톤 진행률 바를 다른 컴포넌트로 추출
마일스톤 진행률 바가 NewIssueSideBarItem 컴포넌트에 하드코딩되어있으니
너무 가독성이 안좋아져서 이를 다른 컴포넌트로 추출함
