+++ client/src/components/newIssue/NewIssueSideBarItem.jsx
@@ -6,6 +6,7 @@ import { IssuesContext } from '../../stores/IssueStore';
 
 import NewIssueSideBarItemTitle from './NewIssueSideBarItemTitle';
 import NewIssueSideBarItemDropdown from './NewIssueSideBarItemDropdown';
+import MilestoneProgressBar from './MilestoneProgressBar';
 
 const styles = {
   layout: {
@@ -23,35 +24,15 @@ const styles = {
       color: 'blue',
     },
   },
-  myProgress: {
-    height: '10px',
-    backgroundColor: '#4CAF50',
-  },
-  myBar: {
-    width: '100%',
-    backgroundColor: '#ddd',
-  },
 };
 
 export default function NewIssueSideBarItem({
   title, defaultMessage, dropdownItems, assigned, setAssigned, author,
 }) {
-  let issues;
-  if (title === 'Milestone') {
-    issues = useContext(IssuesContext).issues;
-  }
-
   const [isAction, toggleAction] = useState(false);
 
   const handleAssignButton = () => {
     toggleAction(!isAction);
-  };
-
-  const progressPercentage = (milestoneId) => {
-    const issuesWithMilestone = issues.filter((issue) => issue.milestoneId === +milestoneId);
-    const closedCount = issuesWithMilestone.filter((e) => e.isClosed === 1).length;
-
-    return (closedCount / issuesWithMilestone.length) * 100;
   };
 
   const assignMyself = () => {
@@ -77,21 +58,11 @@ export default function NewIssueSideBarItem({
         {assigned.length === 0
           ? title === 'Assignees'
             ? (
-              <>
-                {defaultMessage}
-                <button type="button" css={styles.selfAssignButton} onClick={assignMyself}>assign yourself</button>
-              </>
+              <button type="button" css={styles.selfAssignButton} onClick={assignMyself}>{defaultMessage}</button>
             )
             : defaultMessage
           : title === 'Milestone'
-            ? assigned.map((element) => (
-              <div css={styles.item}>
-                <div css={styles.myBar}>
-                  <div css={{ ...styles.myProgress, width: `${progressPercentage(element.id)}%` }} />
-                </div>
-                {element.name}
-              </div>
-            ))
+            ? <MilestoneProgressBar assignedMilestone={assigned[0]} />
             : assigned.map(({ name, color }) => (
               <div css={{ ...styles.item, background: color }}>
                 {name}
